move to printer directory
